# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: VetCRM Admin API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /orders:
    get:
      operationId: listOrders
      tags: [orders]
      summary: List all orders
      parameters:
        - name: status
          in: query
          schema:
            type: string
        - name: search
          in: query
          schema:
            type: string
        - name: vetId
          in: query
          schema:
            type: ["number", "null"]
        - name: page
          in: query
          schema:
            type: number
        - name: limit
          in: query
          schema:
            type: number
      responses:
        "200":
          description: List of orders
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OrderList"
    post:
      operationId: createOrder
      tags: [orders]
      summary: Create a new order
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OrderInput"
      responses:
        "201":
          description: Created order
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
  /orders/{id}:
    get:
      operationId: getOrder
      tags: [orders]
      summary: Get order by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: number
      responses:
        "200":
          description: Order
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
        "404":
          description: Not found
    patch:
      operationId: updateOrder
      tags: [orders]
      summary: Update order
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: number
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OrderUpdate"
      responses:
        "200":
          description: Updated order
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
    delete:
      operationId: deleteOrder
      tags: [orders]
      summary: Delete order
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: number
      responses:
        "204":
          description: Deleted
  /orders/{id}/status:
    patch:
      operationId: updateOrderStatus
      tags: [orders]
      summary: Update order status
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: number
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OrderStatusUpdate"
      responses:
        "200":
          description: Updated order
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
  /orders/{id}/assign:
    patch:
      operationId: assignOrderVet
      tags: [orders]
      summary: Assign vet to order
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: number
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OrderAssign"
      responses:
        "200":
          description: Updated order
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"

  # ── VETS ───────────────────────────────────────────────────────────
...
  /pets:
    get:
      operationId: listPets
      tags: [pets]
      summary: List all pets
      parameters:
        - name: search
          in: query
          schema:
            type: string
        - name: species
          in: query
          schema:
            type: string
        - name: clientId
          in: query
          schema:
            type: ["number", "null"]
      responses:
        "200":
          description: List of pets
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Pet"
    post:
      operationId: createPet
      tags: [pets]
      summary: Create pet
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PetInput"
      responses:
        "201":
          description: Created pet
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Pet"
  /pets/{id}:
    get:
      operationId: getPet
      tags: [pets]
      summary: Get pet by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: number
      responses:
        "200":
          description: Pet
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Pet"
        "404":
          description: Not found
    patch:
      operationId: updatePet
      tags: [pets]
      summary: Update pet
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: number
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PetUpdate"
      responses:
        "200":
          description: Updated pet
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Pet"
    delete:
      operationId: deletePet
      tags: [pets]
      summary: Delete pet
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: number
      responses:
        "204":
          description: Deleted
  /pets/{id}/medical-records:
    get:
      operationId: getPetMedicalRecords
      tags: [pets]
      summary: Get medical records for a pet
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: number
      responses:
        "200":
          description: Medical records
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/MedicalRecord"
    post:
      operationId: createMedicalRecord
      tags: [pets]
      summary: Create medical record
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: number
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MedicalRecordInput"
      responses:
        "201":
          description: Created record
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MedicalRecord"

  # ── SERVICES ───────────────────────────────────────────────────────
...
components:
  schemas:
...
    Order:
      type: object
      required:
        - id
        - clientId
        - serviceId
        - address
        - status
        - totalPrice
        - createdAt
        - updatedAt
      properties:
        id:          { type: number }
        clientId:    { type: number }
        vetId:       { type: ["number", "null"] }
        petId:       { type: ["number", "null"] }
        serviceId:   { type: number }
        address:     { type: string }
        scheduledAt: { type: ["string", "null"] }
        status:      { type: string }
        notes:       { type: ["string", "null"] }
        totalPrice:  { type: number }
        createdAt:   { type: string }
        updatedAt:   { type: string }
        # Joined fields
        clientName:   { type: ["string", "null"] }
        clientPhone:  { type: ["string", "null"] }
        vetName:      { type: ["string", "null"] }
        petName:      { type: ["string", "null"] }
        petSpecies:   { type: ["string", "null"] }
        petBreed:     { type: ["string", "null"] }
        serviceName:  { type: ["string", "null"] }
        serviceIcon:  { type: ["string", "null"] }
        priceFrom:    { type: ["number", "null"] }
        priceTo:      { type: ["number", "null"] }
    OrderList:
      type: object
      required: [orders, total]
      properties:
        orders:
          type: array
          items:
            $ref: "#/components/schemas/Order"
        total: { type: number }
    OrderInput:
      type: object
      required: [clientId, serviceId, address, totalPrice]
      properties:
        clientId:    { type: number }
        vetId:       { type: ["number", "null"] }
        petId:       { type: ["number", "null"] }
        serviceId:   { type: number }
        address:     { type: string }
        scheduledAt: { type: string }
        notes:       { type: string }
        totalPrice:  { type: number }
    OrderUpdate:
      type: object
      properties:
        vetId:       { type: ["number", "null"] }
        petId:       { type: ["number", "null"] }
        address:     { type: string }
        scheduledAt: { type: string }
        notes:       { type: string }
        totalPrice:  { type: number }
    OrderStatusUpdate:
      type: object
      required: [status]
      properties:
        status:  { type: string }
        comment: { type: string }
    OrderAssign:
      type: object
      required: [vetId]
      properties:
        vetId: { type: number }

    # ── VET ────────────────────────────────────────────────────────
...
    Pet:
      type: object
      required: [id, name, species, clientId, createdAt]
      properties:
        id:           { type: number }
        name:         { type: string }
        species:      { type: string }
        breed:        { type: ["string", "null"] }
        age:          { type: ["number", "null"] }
        weight:       { type: ["number", "null"] }
        color:        { type: ["string", "null"] }
        notes:        { type: ["string", "null"] }
        clientId:     { type: number }
        clientName:   { type: ["string", "null"] }
        createdAt:    { type: string }
    PetInput:
      type: object
      required: [name, species, clientId]
      properties:
        name:      { type: string }
        species:   { type: string }
        breed:     { type: string }
        age:       { type: number }
        weight:    { type: number }
        color:     { type: string }
        notes:     { type: string }
        clientId:  { type: number }
    PetUpdate:
      type: object
      properties:
        name:    { type: string }
        species: { type: string }
        breed:   { type: string }
        age:     { type: number }
        weight:  { type: number }
        color:   { type: string }
        notes:   { type: string }
    MedicalRecord:
      type: object
      required: [id, petId, diagnosis, createdAt]
      properties:
        id:          { type: number }
        petId:       { type: number }
        orderId:     { type: ["number", "null"] }
        vetId:       { type: ["number", "null"] }
        vetName:     { type: ["string", "null"] }
        diagnosis:   { type: string }
        treatment:   { type: ["string", "null"] }
        prescription: { type: ["string", "null"] }
        nextVisit:    { type: ["string", "null"] }
        notes:        { type: ["string", "null"] }
        createdAt:    { type: string }
    MedicalRecordInput:
      type: object
      required: [diagnosis]
      properties:
        orderId:      { type: number }
        vetId:        { type: number }
        diagnosis:    { type: string }
        treatment:    { type: string }
        prescription: { type: string }
        nextVisit:    { type: string }
        notes:        { type: string }

    # ── SERVICE ────────────────────────────────────────────────────
...