# lib/api-zod/src/generated/types/medicalRecord.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * VetCRM Admin API
 * OpenAPI spec version: 0.1.0
 */

export interface MedicalRecord {
  id: number;
  petId: number;
  /** @nullable */
  orderId?: number | null;
  /** @nullable */
  vetId?: number | null;
  /** @nullable */
  vetName?: string | null;
  diagnosis: string;
  /** @nullable */
  treatment?: string | null;
  /** @nullable */
  prescription?: string | null;
  /** @nullable */
  nextVisit?: string | null;
  /** @nullable */
  notes?: string | null;
  createdAt: string;
}