# lib/api-zod/src/generated/types/order.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * VetCRM Admin API
 * OpenAPI spec version: 0.1.0
 */

export interface Order {
  id: number;
  clientId: number;
  /** @nullable */
  vetId?: number | null;
  /** @nullable */
  petId?: number | null;
  serviceId: number;
  address: string;
  /** @nullable */
  scheduledAt?: string | null;
  status: string;
  /** @nullable */
  notes?: string | null;
  totalPrice: number;
  createdAt: string;
  updatedAt: string;
  /** @nullable */
  clientName?: string | null;
  /** @nullable */
  clientPhone?: string | null;
  /** @nullable */
  vetName?: string | null;
  /** @nullable */
  petName?: string | null;
  /** @nullable */
  petSpecies?: string | null;
  /** @nullable */
  petBreed?: string | null;
  /** @nullable */
  serviceName?: string | null;
  /** @nullable */
  serviceIcon?: string | null;
  /** @nullable */
  priceFrom?: number | null;
  /** @nullable */
  priceTo?: number | null;
}