# lib/api-zod/src/generated/types/orderList.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * VetCRM Admin API
 * OpenAPI spec version: 0.1.0
 */
import type { Order } from './order';

export interface OrderList {
  orders: Order[];
  total: number;
}