# lib/api-zod/src/generated/types/pet.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * VetCRM Admin API
 * OpenAPI spec version: 0.1.0
 */

export interface Pet {
  id: number;
  name: string;
  species: string;
  /** @nullable */
  breed?: string | null;
  /** @nullable */
  age?: number | null;
  /** @nullable */
  weight?: number | null;
  /** @nullable */
  color?: string | null;
  /** @nullable */
  notes?: string | null;
  clientId: number;
  /** @nullable */
  clientName?: string | null;
  createdAt: string;
}