# lib/api-zod/src/generated/types/petUpdate.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * VetCRM Admin API
 * OpenAPI spec version: 0.1.0
 */

export interface PetUpdate {
  name?: string;
  species?: string;
  breed?: string;
  age?: number;
  weight?: number;
  color?: string;
  notes?: string;
}